requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
django==3.2.9
Cython==0.29.28
easydict==1.10
grpcio==1.51.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.6.0.66
pandas==1.3.5
peewee==3.15.4
Pillow
protobuf==3.20.3
psutil==5.9.4

ipython
PyYAML==6.0
seaborn

requests==2.28.1
requests-oauthlib==1.3.1
scipy
six==1.16.0
smart-open==6.2.0
sqlparse==0.4.3
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop
torch==1.8.0
torchvision==0.9.0
tqdm==4.64.1
openvino==2023.1.0
openvino-dev==2023.1.0
onnx==1.13.0
nncf==2.7.0
